fix: project page root moved
