+++ web/app/[workspaceSlug]/(projects)/projects/(list)/page.tsx
@@ -1,11 +1,4 @@
-"use client";
-import { useParams } from "next/navigation";
-import { useWorkspace } from "@/hooks/store";
-import ProjectsPageRoot from "@/plane-web/components/projects/page";
+import ProjectPageRoot from "@/plane-web/components/projects/page";
 
-const ProjectsPage = () => {
-  const { currentWorkspace } = useWorkspace();
-  const { workspaceSlug } = useParams();
-  return <ProjectsPageRoot currentWorkspace={currentWorkspace} workspaceSlug={workspaceSlug.toString()} />;
-};
+const ProjectsPage = () => <ProjectPageRoot />;
 export default ProjectsPage;
